Set index in pandas dataframe in place when loading Feather databases.
Set index in pandas dataframe in place when loading Feather databases
as this avoid copying the whole dataframe. This makes loading Feather
database faster and avoids allocating memory for a second copy of the
rankings database, just for setting the index. When the whole Feather
database is loaded in the filesystem cache, the conversion from Feather
to a pandas dataframe takes half the time compared with inplace=False.
For example, loading a big region-based database goes from 90 seconds
to 40 seconds (Feather database loaded in filesystem cache first).